Fix most -Weverything warnings.

+++ examples/benchmark.h
@@ -38,7 +38,7 @@ double benchmark(F op) {
       break;
     }
 
-    long next_iterations = std::ceil((min_time_s * 2) / time_per_iteration_s);
+    long next_iterations = static_cast<long>(std::ceil((min_time_s * 2) / time_per_iteration_s));
     iterations = std::min(std::max(next_iterations, iterations), iterations * 10);
   }
   return time_per_iteration_s;
