+++ test/performance.cpp
@@ -31,7 +31,8 @@ TEST(performance_dense_copy) {
 
   dense_array<int, 3> c(b.shape());
   double memcpy_time = benchmark([&] {
-    std::memcpy(&c(0, 0, 0), &a(0, 0, 0), a.size() * sizeof(int));
+    std::memcpy(&c(0, 0, 0), &a(0, 0, 0),
+                static_cast<size_t>(a.size()) * sizeof(int));
   });
   check_pattern(c);
 
@@ -54,7 +55,7 @@ TEST(performance_dense_cropped_copy) {
     for (int z : c.z()) {
       for (int y : c.y()) {
 	       std::memcpy(&c(c.x().min(), y, z), &a(c.x().min(), y, z),
-	                   c.x().extent() * sizeof(int));
+	                   static_cast<size_t>(c.x().extent()) * sizeof(int));
       }
     }
   });
